Division fixed to handle warlpiri corner cases, some changes to "look" of multiplication
	modified:   test-warlpiri-division.jl
	modified:   test-warlpiri.jl
	modified:   ../unum-constants.jl
	modified:   ../unum-division.jl
	modified:   ../unum-int64op.jl
	modified:   ../unum-multiplication.jl

+++ unum-constants.jl
@@ -42,7 +42,7 @@ end
 export nan, nan!
 
 import Base.inf
-function inf{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask = z16)
+function inf{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask::Uint16 = z16)
   esize::Uint16 = 1 << ESS - 1
   fsize::Uint16 = 1 << FSS - 1
   Unum{ESS,FSS}(fsize, esize, signmask, fillbits(-(1 << FSS), __frac_cells(FSS)), mask(1 << ESS))
@@ -52,13 +52,13 @@ neg_inf{ESS,FSS}(T::Type{Unum{ESS,FSS}}) = inf(T, UNUM_SIGN_MASK)
 export inf, pos_inf, neg_inf
 
 #mmr and sss - "more than maxreal" and "smaller than subnormal", ubits for very small things.
-function mmr{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask = z16)
+function mmr{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask::Uint16 = z16)
   esize   ::Uint16 = 1 << ESS - 1
   fsize   ::Uint16 = 1 << FSS - 1
   max_exp ::Uint64 = 1 << (esize + 1) - 1
   Unum{ESS,FSS}(fsize, esize, signmask | UNUM_UBIT_MASK, fillbits(1 - (1 << FSS), __frac_cells(FSS)), max_exp)
 end
-function sss{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask = z16)
+function sss{ESS,FSS}(::Type{Unum{ESS,FSS}}, signmask::Uint16 = z16)
   esize::Uint16 = 1 << ESS - 1
   fsize::Uint16 = 1 << FSS - 1
   Unum{ESS,FSS}(fsize, esize, signmask | UNUM_UBIT_MASK, superzero(__frac_cells(FSS)), z64)
